[BugFix]  `get_json_string` returns NULL (#7243)
This PR fixes the problem that the get_json_string returns NULL when a JSON array is an input. The previous implementation of get_json_xx is implemented with simdjson directly, which is confusing and hard to maintain. The new implement unifies the implement of get_json_xx and the functions of JsonValue and simplifies the get_json_xx.