[Enhancement] Estimate scan row bytes dynamically in time (#7202)
The current mechanism of estimating the scanning column bytes is to directly give a fixed estimated value. If the estimated value deviates greatly from the actual value, the parallelism will be too low.
For example, with a Bitmap estimate of 1MB, query_mem_limit of 32GB would limit the parallelism to 1.

Therefore, this PR adds the strategy of "adaptive row bytes estimation".
During the scan tasks running, the estimated_row_bytes is adjusted in real time according to the read chunk bytes, and then the parallelism is adjusted.