[BugFix] Use independent thread to write journals by batch asynchronously (#6988)
We have found some severe performance issues described in #4327.
The main reason is that each thread that writes one journal needs to call EditLog.logEdit(), a synchronized method that holds a global lock. As a result, this function becomes the bottleneck of the journal writing pipeline.
We can use an independent thread called JournalWriter to write journals by batch asynchronously. Each thread that needs to write a log can put the log in a blocking queue, while JournalWriter constantly gets as many logs as possible from the queue and write them all to BDB in one transaction. After committing, JournalWriter will notify the caller thread for consistency.