[Cherry-pick][Bugfix] Fix bug in some case failed to set DefineExpr for materialized view (#7427)
* Fix bug in some case failed to set DefineExpr for materialized view. (#7371)

If the user's table contains uppercase columns, then defineExpr may be processed in two cases when building a materialized view. One is that Log is processed through MVColumn, which is no problem. The second type of Image is directly compared to Column. This processing is problematic, which will cause defineExpr to be set to null, which will affect import and query.