[BugFix] Fix load jobs hang with error: current running txns on db xxx is 100, larger than limit 100 (#7414)
After transferring the master, the master address recorded in be is still the address of the old master(the time before it reaches the new master's heartbeat). The txnCommit rpc executed on non-master fe will cause some metadata inconsistency issues (described in #7350). So we should reject those request if current node is not master.