[BugFix]: Load txn commit failed while primary table using persistent index (#7450)
If the primary table using persistent index, we will create a `index.l0.x.x` to save the snapshot or log. However, after the first data load is successful, the second data load may fail. 

The reason is when we dump snapshot of the `hash_map` in memory, we use function `dump_bound()` to evaluate the snapshot file size. But when the map is empty, the return value of function `dump_bound()`  will be set as larger than sizeof(uint64_t) but the snapshot file only write `size_(uint64_t)`. It will cause the meta data error which leads to deserialize `index.l0.x.x` failed.

So the return value of function `dump_bound` is set to `sizeof(size_t)` when the `_map` is empty