[BugFix] use new parser in meta replay & restore (#7481)
The root cause is a compatibility issue between the old parser and the new one.
When parsing multiple statements with comments, the old parser will get one statement while the new parser will get two, the first of which is an empty statement. Thus if we store the index from the result that is parsed from a new parser, then try to get the statement using results that are parsed from an old parser, we'll get an index error.
I'm replacing the old parser with the new one to solve this part.

(cherry picked from commit 91bb9c5e627ddba127bf63577812bdd95ef51080)