[BugFix] Fix broker exception the fd is not owned by client null (#7773)
Broker has control plane heartbeat service. It may be a timeout when the pressure is high. Then it will be removed from clientContexts and make the data plane throw FD is not owned by client null exception.
So that we update the heartbeat timestamp in both the control plane & data plane to avoid this problem.

(cherry picked from commit 6c56ddc5083bbf28263c7178ed1689654adef9cf)