set defineExpr use new parser (#7684)
DefineExpr was parsed with the old parser before, but the new one was used when building the table, so there may be a problem of inconsistent statements. So we switch to the new one to cover all the old parsing cases.

(cherry picked from commit 46741ee6a0ab3b377b4302e24b95c5eb8fbd2007)