[Bugfix] Prevent JVM from handling signals (#6929) (#7841)
When the signal SIGINT or SIGTERM coming, the process BE captures it and performs graceful exit operations by order. For example, when a SIGTERM coming,

The process captures SIGTERM to make start_be() can be returned.
The thread TaskWorkerPool, which uses StarRocksMetrics::instance, is desctructed, when the function start_be() returns.
__run_exit_handlers desctruct all the static variables including StarRocksMetrics::instance.
However, JVM overwrites the handlers of SIGINT and SIGTERM. Then, the sequential graceful exit operations are breaked. start_be() has no chance to return, so StarRocksMetrics::instance will be desctructed while The thread TaskWorkerPool is running.