[Enhancement] Add ut for compaction with empty rowset (#7817)
It's suspected that compaction may have some bug related to empty rowsets, so this PR adds some UT for this scenario. Although no bug is found.

This PR also adds checks to check input strings' lengths and easily identify bugs. And as a result, a bug in TestGlobalDictNotSuperSet and TestGlobalDictNoLocalDict was found.