[BugFix] fix replica meta falsely deleted right after clone finished (#7905)
The main reason causes this problem is that FE handles tablet report tasks and clone
tasks concurrently, with specific timing, this will happen. So we add a new state
`deferReplicaDeleteToNextReport` for `Replica`, default to true. When FE meets a replica
that only existed in FE's meta, not in the tablet report, it will check
`deferReplicaDeleteToNextReport` and defer the meta delete till the next report of the BE.
In the following situation, a normally cloned replica could be falsely deleted:
1. BE X generates a tablet report and sends it to FE
2. FE creates a clone task(for balance or repair) and a new replica on BE X,
     so the corresponding tablet is not included in the report sent above.
3. BE X finishes the clone and then FE will receive the message and set the state
     of the new replica to NORMAL
4. FE processes the tablet report from step 1 and finds that BE X doesn't report
    the tablet info corresponding to the newly created replica, so it will delete
    the replica from its meta
5. On the next tablet report of BE X which will include the tablet info, FE finds
     that the tablet reported by BE X doesn't exist in its meta, so it will send a
     a request asking BE X to delete the newly cloned replica physically.

(cherry picked from commit 03070130c8dd1d38d50f8d60a3f0d07d3c60f944)