[BugFix] Fix the check the state of opening sink
To improve the performance, the sink will call try_open() in prepare() of OlapTableSinkOperator.
And then call open_wait() in the push_chunk() to wait the sink opened.
If the chunk is empty, the pull_chunk() have been not called.
So we should call open_wait() in set_finishing() to wait the sink opened.
Otherwise, the close() and open() will be overlapped and cause a error.