[BugFix] Fix bug that thrift server daemon exits quietly (#7974)
Rewrite the execute() function of org.apache.thrift.server.TThreadPoolServer: We do not kill the server for any exception thrown by ExecutorService, just close the connection and print the error log. But TThreadPoolServer will kill the server, and any new connection will not be processed, it will cause the connections saturate the tcp backlog.

Because there are some private properties in TThreadPoolServer. So we create a new implementation of TServer, and copy the code of TThreadPoolServer, and just modify the error process of the execute function.

(cherry picked from commit 3e49ab438f579be7dd88b824cd8208ed5cd36307)