[BugFix] Notify transaction status visible to caller after write log (#8421)
When the txn state of stream load changes to be visible, it returns to the caller that the stream load has been completed. In fact, the visible log has not been written, at this time, if the query is executed from the follower, it will miss this batch of data. So the current change makes after the visible log is written, it will return to the caller that the stream load is completed.