[BugFix] Fix potential dangling pointer of TabletSchema in Segment
For lake tablet, the raw pointer of TabletSchema in Segment may point
to a memory location that has been deleted, if the cached TabletSchema
has been evicted from the in-memory cache of lake::TabletManager.

Use a std::shared_ptr to manage the ownership of TabletSchema in Segment
to fix the issue.