[BugFix] Persist `LoadStatistic` in `EtlStatus` (#8689)
The main reason is that LoadJob did not persist LoadStatistic in the process of either replaying or snapshot. Unfortunately, there is no obvious way to add new metadata since we persist the LoadJob class in a hard-coded way. After due consideration, we serialized LoadStatics in JSON for the convenience of adding/deleting fields in the future and hide this JSON in a deprecated map in EtlStatus.

(cherry picked from commit f931637a65db4d5fb10c77a937fb5e662ed9f098)