[Enhancement] Reject create table request if checkPersistentIndex failed (#9021)
PersistentIndex is available in branch-2.3, but there are many limitations in key columns. For example, key column is not supported as varchar(char) type, and the length of key columns cannot exceed 64 bytes.

In the previous implementation, if we create a primary key table enable PersistentIndex but the restrictions are not met, we will use PrimaryIndex in memory but not PersistentIndex, and primary key table will be create successfully. This may be a bit confusing to the user.

So we will reject the create table request directly if we find that the restrictions are not met.