[Feature] Support prepare transaction stream load interface for 2PC (#8403)
StarRocks implements the /api/transaction/prepare interface to put the current transaction to be PREPARED state which temporarily persists changes. After preparing a transaction, the user can continue to commit or rollback the transaction.

The transaction preparation interface helps users achieve a more robust cross-system two-phase commit. After the transaction preparation is successful, even if StarRocks crashes, the user can continue to perform the commit or rollback this transaction after the system is recovered.