[BugFix] Primary key length is inconsistent with be (#9148)
PersistentIndex is available in version 2.3, but there are many limitations in key columns. For example, key column is not supported as varchar(char) type, and the length of key columns cannot exceed 64 bytes. However, the calculate logic in fe is inconsistent with be right now(DATE/DATETIME), so we may reject some create request actually we can do.