Fix stats npe with unknown column when query with low card optimize (#9173)
When low cardinality query optimization is turned on, there may be columns that do not exist in the table, which will lead to NPE when the statistics sql is built
During low cardinality optimization, a new type of columnRef of type agg+column_name will be constructed and filled into LogicalMetaScan. resulting in columns that are not present in the table. Currently modified to column name to get statistics