[Enhancement] rename `currentKey` to `nextKey` in `BDBJournalCursor` (#9376)
A long long time ago in a galaxy far far away, there was a litter variable called currentKey that always referred to the key of the journal to be replay instead of the current one, unless when inside the next() function. This name has successfully confused some aliens that arrive ten years later, and will be changed to nextKey which they all agreed is much more accurate.