[BugFix] shouldn't close file system when broker encounters excecption (#9507)
In current implementation, broker will close file system when it encounters exception.
Suppose there are two threads, threads A is in the middle of reading process, thread B
is opening the file and it finds the file doesn't exist, then broker will throw an exception
and close the file system. However, thread A is still using the file system to read files,
close the file system will also lead to thread A abort reading.