[BugFix] null date and datetime value for spark load is not consistent with broker/stream load
For broker/stream load, null value for date and datetime is DefaultValueGenerator<DateValue/TimestampValue>,
however, for spark load, it is not so.
So when we create a aggregate table, we first perform broker load, and then perform spark load on the same data
(here assuming the data has null value), and query by select count(*) from table, the result will contain two null value.

In this pr, we let the null date and datetime value for spark load to be consistent with broker/stream load

Signed-off-by: xyz <a997647204@gmail.com>