[BugFix] fix `SHOW PROC /colocation_group' after erase db with same name (#9604)
In my previous implementation, when executing SHOW PROC /colocation_group, there will be a star next to the table id if the ID of the database / table is in CatalogRecyleBin.
If we create & drop databases with the same name twice, the first database will lose its Database ID from CatalogRecycleBin. As a result, it will look like a normal table in the output.
To fix this, now we only validate if the ID of the database or table in GlobalStateMgr.