[Feature] Add a case sensitive flag to hdfs scan node to indicate whether the hive table column names are case sensitive or not.
By default, presto/hive/impala are column name case in-sensitive.
But spark and iceberg can be configured to be column name case sensitive or not.
So when reading extern tables from different data source, we need to know
whether the column name are case sensitive, if not, we shold compare the
column names ignoring case.