[Enhancement][Lake] Avoid sending unnecessary publish version tasks
This patch is necessary for supporting schema changes on lake table.

For local tables(i.e., OlapTables), you can send PublishVersionRequest
to tablets of a shadow materialized index, even if the transaction does
not write any data to the shadow materialized index. But for LakeTable,
sending PublishVersionRequest to tablets of a shadow materialized index
that has not been written will result in errors.

This patch also removed two unused member variables(rollupIndexId and
rollupFinishedVersion) from MaterializedIndex.