[cherry-pick][branch-2.3][Feature] Add a case sensitive flag to hdfs scan node to indicate whether data source is column name case sensitive (#9744)
By default, presto/hive/impala is column name case insensitive. But spark and iceberg can be configured to be
column name case sensitive or not. So when reading extern tables from different data sources,
we need to know whether the column name is case sensitive; if not, we should match the column names by ignoring the case.