Fix Unstable UT SchemaChangeJobV2Test.testSchemaChange1 (#10096)
This test is actually a schemachange process test, but it relies on the operation of executing the Alter Task in the background. This operation may be executed in advance and cause a task error. And the logic of writing UT before is actually a FINISH operation implemented by itself, which is actually not good. Here it is changed to just wait for FINISH to complete the operation naturally.

(cherry picked from commit 22bf5e86139784d3d8cf887b53d16d1d50f63356)