[BugFix] Replayer quit on exception if environment is invalid (#10152)
When the disk is full, or the permission of the BDB directory has been mistakenly changed, the replayer will catch the exception and retry in vain.
This PR will handle EnvironmentFailureException with great care, and will raise a JournalInconsistentException to replayer when the current replicated environment is invalid so that the whole process will quit.

(cherry picked from commit e844bc5eb4b9d61654bb8c4e7850d55d9306bff7)