[Enhancement] Leader transfer should not impact read ops (#9106)
When follower is transferring to leader, the meta will be set to unreadable. All read ops will be forward to the old leader, but the old leader is not available (crashed or stuck), so the read will fail. In fact the leader's meta is the readable, setting it to unreadable is unreasonable.