[Enhancement] Rebuild persistent index if load previous index failed (#10244)
We will try to build a persistent index according to the index file. If some error occurs during the loading index, we will return fail directly. But we can rebuild the persistent index according to tablet data. So we will try to rebuild the persistent index when some error occurs.

In addition, we will support varlen key for the PrimaryKey table with a persistent index from version 2.4. However, if we create a PrimaryKey table with a persistent index whose key length is less than 64 Bytes in version 2.4, we can't rollback to version 2.3 directly because the meta structure is changed in version 2.4.

This pr also supports rollback from version 2.4 to version 2.3 for PrimaryKey table with a persistent index.