Modify the implementation of the parse root path property (#10450)
The reason is that the path is parsed into property and ignored directly, resulting in one less disk and triggering a large number of clone task.

the usage such as /data,50 has been removed long long ago. So if we get only one value from parsing property, we will report error directly.