[BugFix] Allow select from view when user is only granted on view (#10508)
When checking the privileges of a query statement, instead of checking the select privileges of all the relative tables, we traversal all nodes and handle them separately. As to view, the operation will be allowed when the user has the privilege to the view itself, otherwise, it will keep looking for select privileges on all the base tables as it used to be.