[Enhancement] Make the be reboot time display more accurate (#10335)
The previous be's restart time is not accurate enough.
Before that, whether fe updates the lastStartTime of be depends on the alive state of be recorded by fe itself.
When the alive state of be is false and the be heartbeat is received again, fe updates the lastStartTime of be.
If fe dose not receive be's heartbeat 3 times, fe will set the alive stale of be to false. So if be reboot quickly, the alive state of be will not be set to false, and the lastStartTime of will not be updated
This PR adds a field reboot time in be, when be reports a heartbeat, reboot time will be sent to fe to determine if lastStartTime of be needs to be updated.