[BugFix] abort insert into txn if filter data in strict mode
The default mode of insertion is strict, and any wrong farmot data will abort the
insertion. In the current implementation, when we abort an insertion.
it does not abort the txn corresponding to that insertion.

In this pr, if we encounter invalid data in a strict mode insertion, we will abort the txn.

Signed-off-by: xyz <a997647204@gmail.com>