[BugFix] Ranges may be lost when dynamic partitions are created. (#10678)
If the partition type changes from the month dimension to the week dimension. Old dynamic partitioning for [(' 2022-08-01 00:00:00), (' 2022-09-01 00:00:00))
In this case, the new dynamic partition of the week dimension [('2022-08-29 00:00:00'), ('2022-09-05 00:00:00')) fails to be created. The next partition created successfully will be [('2022-09-05 00:00:00'), ('2022-09-12 00:00:00')). No partition can be found for the data from September 1 to September 4. Therefore, I added a dynamic clipping partition in this area. When this situation is detected, I clipped the starting position of the partition to create a partition for [('2022-09-01 00:00:00'), ('2022-09-05 00:00:00')).