[BugFix] Set initial value type to size_t to avoid overflow (#10741)
The result type of function std::accumulate is the same as the initial value, and the default type of initial value is int32. So if the total l1 file size is more than 2GB, the following code will cause an overflow exception.
```
        const auto l1_kv_pairs_size = std::accumulate(std::next(_l1->_shards.begin(), l1_shard_offset),
                                                      std::next(_l1->_shards.begin(), l1_shard_offset + l1_shard_size),
                                                      0, [](size_t s, const auto& e) { return s + e.size; });
        const auto l1_kv_pairs_usage = std::accumulate(std::next(_l1->_shards.begin(), l1_shard_offset),
                                                       std::next(_l1->_shards.begin(), l1_shard_offset + l1_shard_size),
                                                       0, [](size_t s, const auto& e) { return s + e.bytes; });
```
So we set the initial value type to size_t to avoid overflow

(cherry picked from commit 5288aaa6c8df9578c8512ec928198f6a9ef0f5a6)