[BugFix] abort insert into txn if filter data in strict mode (#10657)
The default insertion mode is strict, and any wrong format data will abort the
insertion. In the current implementation, when we abort an insertion.
it does not abort the txn corresponding to that insertion.

In this pr, if we encounter invalid data in a strict mode insertion, we will abort the txn.