[BugFix]Fix colocate state inaccuracy bug (#10642)
This PR #9354 added the check logic that if the tablet has been in the tablet scheduler queue, then do not check the health status. But the isGroupStable will remain true if all the tablets are in the tablet scheduler queue, which makes the status inaccurate. Set isGroupStable to false if there is tablet in the tablet scheduler queue.

(cherry picked from commit 923f5f05671b9d2018412880b16112da9f4b1dc9)