[Enhancement] Add mem statistics for rowset (#11057)
Currently it is too hard to troubleshoot metadata memory problem, and more detailed statistics are required.
Add mem statistics for segment for rowset metadata.

RowsetMeta may be modifyed after create, so it may be not inconsistent at construct and destruct using `_rowset_meta_pb->SpaceUsedLong`, So we add one item to record the mem usage. This method will have a certain deviation, but it can ensure that the statistical error will not accumulate.

Remove the default constructor of `RowsetMeta` for easy mem statistics.

Signed-off-by: trueeyu <lxhhust350@qq.com>