[BugFix] fix export close file fail bug
If there is an error occurs in export, such as PlainTextBuilder init fail because of No CSV converter for type OBJECT,
export will return from BE to FE immediately, and FE will clean the temp dir.
At the same time, BE will close the temp file in temp dir.
if unfortunately, the temporary file to be closed has already been deleted, it will report that the close file is deleted error msgs.

In this pr, we will return to FE until we close all temp file.

Signed-off-by: xyz <a997647204@gmail.com>