[UT] fix unstable UT BDBEnvironmentTest#testRollbackExceptionOnSetupCluster (#11248)
This test case tries to simulate an unexpected scenario where a master was down before it got ahead of all followers. Normally BDB will cause a `RollbackException`, which we should handle and turn to a `JournalException`. But there is a slight chance that BDB may handle it well and nothing goes wrong. Either way, we think it's reached our expectations.