[BugFix] Add min readable version report (#10694)
When a newly cloned replica is added to Tablet's replica list, it only has a single version V1. Still, when FE executes a query, it may read the tablet with a version Vr < V1. Because V1 > Vr, it's a valid candidate for the tablet's read. If FE chooses this replica for the reading when executing ScanNode, a "version not found" error will occur, and the query will fail. This situation is also possible if version Vr is compacted and GCed, but less likely than the clone situation.

This PR adds minimal readable version information to the Replica object in FE. When the clone finishes or BE reports tablet infos, BE will report the replica's minReadableVersion to FE, and FE can update this value. So when scheduling a replica for a scan, replicas without the specified readable version can be skipped, avoiding the "version not found"/"version is compacted" error.