[Enhancement] raise `JournalInconsistentException` instead of `System.exit()` in `EditLog.loadJournal` (#11249)
Raise JournalInconsistentException instead of calling System.exit() when fatal error occurs in EditLog.loadJournal() and the process must terminate itself. All JournalInconsistentException will be caught and handled in GlobalStateMgr.replayJournalInner, including printing the id of the current journal. Thus, if a journal cannot be loaded by EditLog.loadJournal(), we can find the journal id and skip it.