[BugFix] all partitions have no load data when jsonroot is set (#11145)
To get better performance in scenes where the column is sparse, we introduced #5357.
When the jsonroot is set, and the jsonpath is not set, the _slot_desc_dict is not initiated, leading the error all partitions have no load data being returned.
This PR fixes the problem and simplifies the JsonScanner by removing the _construct_row_in_slot_order. In this way, the row would only be constructed in JSON object order, and the JSON object would only be iterated once.