[Feature] support CONVERT/UNSIGNED for metabase query (#11388)
For the UNSINGED type, we treat it as the same as INTEGER, and for CONVERT we treat it as the same as CAST, so that we can support metabase's metadata query syntax.

In some extreme cases, there may be accuracy problems, but we can encounter such problems and further fix them.

(cherry picked from commit 91af979328a67e9b8fef10deebc3c086b71a7835)