[Feature] new privilege framework: basic authentication (#11447)
Introduce a new privilege framework.
To avoid huge PR which is hard to review and easy to conflict, I'm splitting into several small PR, each implementing an individual aspect of the new privilege framework. This is the first one.
This PR implements basic authentication and supports creating a user with MySQL plain password.