[Enhancement] Optimize profile string memory overhead (#11570)
ProfileManager keep 1000 profile infos of recently executed query,
the average length of profileContent string is about 600KB+ in a daily test
environment, and with some complicated query, this length can go up to
5MB+, which can consume a lot of memory. The profile string has a very good
compression ratio; with the simple gzip test, we can see the memory overhead
reduced to about one-tenth which is a big save, so we use the compression
method to optimize this. And we also added a new dynamic configuration to
control the reserved profile infos.