[BugFix] BE crashed on invalid JSON input with jsonpath (#10838) (#11507)
The JSONReader does not return an error when the input is invalid JSON and with jsonpath, leading to the BE crash.
This PR divides the return error of `JsonFunctions::extract_from_object` as 2 types:
  1. NotFound: The key is not found.
  2. DataQualityError: the input is invalid.
When `JsonFunctions::extract_from_object` returns NotFound, the next key will continue to be looked up in the object. When the DataQualityError is returned, the input is valid, and the JSONReader returns the error directly.