[Enhancement] Change LinkedListMultimap to ArrayListMultimap to better support for getting elements by index (#11774)
When creating a table, FE will first put the tablet to be created into the TabletInvertedIndex and then send the task to BE to create the tablet. And when BE reports the tablet, FE will find the difference between the tablet reported by BE and the tablet in TabletInvertedIndex, at this time, if the bucket of table creation is relatively large, there will be a lot of extra tablet in TabletInvertedIndex. But in processing these extra tablet, FE put it into a LinkedList, and when getting the elements, it uses the get(int index) method, resulting in extremely inefficient. And this processing logic is put into the database write lock, resulting in the database stuck.
To fix this bug, we should use arrayList.