[Feature] New RBAC framework: support alter & drop user (#11732)
Introduce a new privilege framework.
To avoid huge PR which is hard to review and easy to conflict, I'm splitting into several small PR, each implementing an individual aspect of the new privilege framework. This is the fourth one. For previous PRs see
#11447
#11528
#11593

This PR implements alter user and drop user.