[BugFix] Check the nullable property based on slot instead of parquet (#11880)
Fix the BE crash problem when select array from iceberg external table with parquet file format.

As we treats all column slots from hive are nullable, and then create columns based on slots property. If the parquet array columns are required, the field parsed from parquet schema will be not nullable. In this case, if we depend on field property, the nullable column will be converted unexpected.