[BugFix] Fix l1 file size keep growing (#11867)
For #11868:

In branch-2.4, we support the varlen key of persistent index. But the estimate l1 size is not accurate because the following two reasons:

We use the total file size of previous l1 file as the real l1 data size
We ignore the same key in l0 and l1, and calculate size twice
These two reasons will cause l1 file size keep growing but the real data size is far smaller than file size.

For #11721:

_total_kv_pairs_usage of SliceMutableIndex is not update when load snapshot. So the _total_kv_pairs_usage is not accurate after load snapshot which leads to an error in the result of dump_bound(). And it may lead to subsequent metadata errors.

So we will use the read data file size as the snapshot size to avoid potential errors.