[BugFix] Support authentication for StarRocks external table (#11871)
When creating StarRocks external table, properties of user and password are required. But currently the target cluster does't check the password and LOAD_PRIV , so the INSERT INTO will be successful even if the user does not exist, or the user has no LOAD_PRIV on the target cluster. We should fix this problem.

When inserting data into the table on the target cluster, the source cluster will issue multiple RPCs to the target cluster in order, including beginRemoteTxn, commitRemoteTxn, and abortRemoteTxn in FrontendServiceImpl. We only need to
do the check in the first RPC beginRemoteTxn. So authentication information will be carried by the RPC, and checked in the target beginRemoteTxn.