[BugFix] Fix FE restart failed when replaying schema change job (#11992)
This bug arises during the process of finishing schema change job, a partition is added to table just after the table state is change to Normal and before the alter log is written to bdb. This is a common safe problem in our system: bdb log is written outside the db lock.
To fix this bug we should write bdb log inside the db lock. But to reduce the lock occupancy time, we just put the log item to the buffer queue before releasing the db lock, and wait the log to persist outside the db lock.

(cherry picked from commit 27a2691025e300e53c08e3247f1fbf2302a69f86)