fix (#12126)
This bug arises during the process of finishing schema change job, a partition is added to table just after the table state is change to Normal and before the alter log is written to bdb. This is a common safe problem in our system: bdb log is written outside the db lock.
To fix this bug we should write bdb log inside the db lock.