[BugFix] Fix fencing failed when a new leader is elected (#12138)
Because of consistency bug in bdb(may arise when all node restart continuously), the data in epochDB may be discontinuous, so the epochDB.count() + 1 can be an existing value, which will cause epochDb.putNoOverwrite failed. To fix bdb bug is difficulty, so we define the new epoch by the largest value in epochDB + 1.