[BugFix]Fix olap external table meta synchronization bug (#12368)
The tablet of the external OLAP table should not be added to TabletInvertedIndex; it will replace the local tablet if they have the same id, which will cause some inconsistency bugs.
Besides, this PR changes the param from isRestore to updateInvertedIndex, which makes the param more visible. And there is no need to cache the OLAP external table metadata in the non-leader node, so if the current node is not a leader, do nothing.

(cherry picked from commit 316654710fe074daf7ecbfb2c1ba6f5337b3c3a6)

# Conflicts:
#	fe/fe-core/src/main/java/com/starrocks/catalog/LocalTablet.java
#	fe/fe-core/src/main/java/com/starrocks/catalog/OlapTable.java
#	fe/fe-core/src/main/java/com/starrocks/external/starrocks/StarRocksRepository.java
#	fe/fe-core/src/main/java/com/starrocks/lake/LakeTable.java