[BugFix] Fix colocate balancer not handling bad replica case (#12588)
If a replica is marked bad, we need to migrate it and all the other replicas corresponding to the same bucket
index in the colocate group to another backend, but the backend where the bad replica sits on may still be
available, so the backend set won't be changed by ColocateBalancer, so we need to learn this state and update
the backend set to replace the backend that has the bad replica. In order to be in consistent with the current
logic, we return COLOCATE_MISMATCH not REPLICA_MISSING in `getColocateHealthStatus()`.