[BugFix] be_crash_when_using_snappy_for_page_compression (backport #10883)
Problem:
When we use snappy for page compression, be may crash. The reason is that, the original snappy lib does not support multiple discontinue memory for compression. So we reform the input parameter for it, which is SnappySlicesSource. But the member function Peek may return wrong pointer which cause this problem.

Solution:
Make SnappySlicesSource.Peek return the correct pointer.