[Refactor] Add sample size to session var when getting hive partition statistics (#12700)
Fixes #
If a query contain external table without metadata cache, It will get all partitions info from hive metastore to calculate table level statistics even if one partition is queried. This behavior existed before 2.5. But if table has too many partitions such as 500000, it will cause the hive meastore to return failure or no return. This will also cause the db read lock of the query not to be released. So we add session variable hive_partition_stats_sample_size to limit the partition size. The default value is 5000.