[Enhancement]Task Framework supported execution progress (#12655)
This PR is a supplement to th PR （#12292），so that in the case of FE high-availability clusters, the progress can still be queried in the follower nodes. The report worker only reports the progress of loadjob to the master, so the progress of task_runs cannot be read out in the followers by executing the sql (selecting * from information.schema.task_runs)

The PR Starts a leader daemon and check the running task_run jobs every five seconds. If the progress is updated, it will be synchronized to the follower nodes through editlog