Cherry-Pick-2.3][Enhancement] Add more safe lock function to Database (#12864)
Many places call db.writeLock after getting db from catalog, and then write bdb log. But the db may be dropped in the process of writing bdb log, which will cause the NPE error when replay bdb log. To solve this problem, add an exist flag to mark whether db is dropped, and users can only obtain the lock when this flag is true.
This PR only add the new lock functions. Subsequent PRs will change the operational logic of the upper layers.
In the future, only a few scenes can get the lock without checking whether the db is dropped, and the others should call the newly added function to get the lock.