[BugFix] Fix primary key table vertical compaction bug (#12797)
In vertical compaction of PrimaryKey table, The `rowsets_mask_buffer` parameter should correspond to the rowsets because we use `rowsets_mask_buffer` to determine the order in which subsequent segments are read. And if rowset only has one segment, we should push a emtpy RowSourceMask into `rowsets_mask_buffer` but not skip it, otherwise the misalignment may cause some unexpected error.