[BugFix] Fix the page selection bug in late materialization for large columns.
The page selection in late materialization depends on `next_row` context, which only be updated
after the total selection procedure now.
As this selection procedure may contain multiple select operations for large column, the context
cannot be updated in time in the subsequent operations, which causes the unexpected selection result.