[BugFix] Different column types in different chunks (#13078)
The column type like `datetime` is casted from `varchar` in the file scanner. `cast_from_string_to_datetime_fn` would build a column as nullable when there're nulls in the column, while it would build a column as non-nullable when there's no null in the column. So, the column types of different chunks may be different. 
The exchange node gets the chunk schema according to the first chunk it receives, and then parses the chunk according to the schema. In this way, the exchange node cannot handle the variable column type.
This PR corrects the columns returned by the `cast_from_string_to_datetime_fn` according to the actual schema.